Change TTU to the specific time to update
TTU could be defined before these changes as the time since "now" to
the time of renovation, what is obviously different depending on the moment
of the calculation because "now" is not constant. After these changes the
TTU is the specific point in time of renovation, that doesn't depend on any
other thing.

This makes operations using the TTU less vulnerable to side-effects, and
increases their testablility. Some basic tests for next TTU are added.