Redistribute vault test code
Some code in testutils.go that is used by some subpackages of pouch
needs code from vault that makes it increase the size of the
binaries. It was mitigated by adding the testutils tag, so a default
build doesn't include the extra code, but it made it neccesary to add
the tag when running tests.

This change redistributes this code, so the code that increases the
size of the binaries is moved to a different package
(./pkg/vault/test) that can be imported in tests needing it.

Other code that doesn't include so many things has been added to the
vault package in the dummy.go file. This can be used by other packages
by importing the pkg/vault package.